backend/requirements.txt
absl-py==2.3.1
asgiref==3.11.0
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
immutabledict==4.2.2
numpy==2.3.5
ortools==9.14.6206
pandas==2.3.3
protobuf==6.31.1
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
